pyproject.toml
[tool.poetry]
name = "cert-downloader"
version = "0.1.0"
description = ""
authors = ["Richard Greville <richard.greville@opticoreit.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyOpenSSL = "^22.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
